+++ src/errors.js
@@ -2,6 +2,8 @@
 export const ERROR_CODES = {
   MISSING_PARAMETER: 'missing_parameter',
   REMOTE_SERVICE_ERROR: 'remote_service_error',
+  INVALID_STATE: 'invalid_state',
+  NO_SESSION_DATA: 'no_session_data',
   UNKNOWN: 'unknown'
 }
 
@@ -105,5 +107,21 @@ export class SignatureVerificationError extends BlockstackError {
     super({ code: 'signature_verification_failure', message })
     this.message = message
     this.name = 'SignatureVerificationError'
+  }
+}
+
+export class InvalidStateError extends BlockstackError {
+  constructor(message: string) {
+    super({ code: ERROR_CODES.INVALID_STATE, message })
+    this.message = message
+    this.name = 'InvalidStateError'
+  }
+}
+
+export class NoSessionDataError extends BlockstackError {
+  constructor(message: string) {
+    super({ code: ERROR_CODES.INVALID_STATE, message })
+    this.message = message
+    this.name = 'NoSessionDataError'
   }
 }